docs/Observability/examples/flask_instrumentation/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
